Added castle map left and right.
they are defense maps so only one side can score. Is meant that each team can
play say 10m and the goal is to beat the other's score when they swap sides

there are three flags for the defending team to defend. maze, (spiral in the top)
treasury (maze at the bottom) and throne room (center near the respawn with random blocks around)